Azure/requirements.txt
azure-identity
azure-mgmt-compute==17.0.0
azure-mgmt-network==16.0.0
azure-mgmt-resource==15.0.0
azure-mgmt-storage==16.0.0